Subdivide large namespace
