+++ Libraries/dotNetRDF/JsonLd/JsonLdProcessorException.cs
@@ -25,6 +25,7 @@
 */
 
 using System;
+using VDS.RDF.JsonLd.Syntax;
 
 namespace VDS.RDF.JsonLd
 {
