+++ Libraries/dotNetRDF/JsonLd/JsonLdProcessorWarning.cs
@@ -24,6 +24,8 @@
 // </copyright>
 */
 
+using VDS.RDF.JsonLd.Syntax;
+
 namespace VDS.RDF.JsonLd
 {
     /// <summary>
