+++ Libraries/dotNetRDF/JsonLd/JsonLdTermDefinition.cs
@@ -27,6 +27,7 @@
 using System;
 using System.Collections.Generic;
 using Newtonsoft.Json.Linq;
+using VDS.RDF.JsonLd.Syntax;
 
 namespace VDS.RDF.JsonLd
 {
