+++ Libraries/dotNetRDF/JsonLd/Processors/ExpandProcessor.cs
@@ -29,9 +29,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Newtonsoft.Json.Linq;
+using VDS.RDF.JsonLd.Syntax;
 using VDS.RDF.Parsing;
 
-namespace VDS.RDF.JsonLd
+namespace VDS.RDF.JsonLd.Processors
 {
     internal class ExpandProcessor : ProcessorBase
     {
