+++ Libraries/dotNetRDF/JsonLd/Processors/FramingProcessor.cs
@@ -29,8 +29,9 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using VDS.RDF.JsonLd.Syntax;
 
-namespace VDS.RDF.JsonLd
+namespace VDS.RDF.JsonLd.Processors
 {
     internal static class FramingProcessor
     {
