+++ Libraries/dotNetRDF/JsonLd/Processors/FramingState.cs
@@ -27,8 +27,9 @@
 using System;
 using System.Collections.Generic;
 using Newtonsoft.Json.Linq;
+using VDS.RDF.JsonLd.Syntax;
 
-namespace VDS.RDF.JsonLd
+namespace VDS.RDF.JsonLd.Processors
 {
     internal class FramingState
     {
