+++ Libraries/dotNetRDF/JsonLd/Processors/NodeMapGenerator.cs
@@ -26,8 +26,9 @@
 
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using VDS.RDF.JsonLd.Syntax;
 
-namespace VDS.RDF.JsonLd
+namespace VDS.RDF.JsonLd.Processors
 {
     /// <summary>
     /// An implementation of the JSON-LD node map generation algorithm.
