+++ Libraries/dotNetRDF/Writing/JsonLdWriterOptions.cs
@@ -25,6 +25,7 @@
 */
 
 using VDS.RDF.JsonLd;
+using VDS.RDF.JsonLd.Syntax;
 
 namespace VDS.RDF.Writing
 {
